- moved "htconfig.php" to "config/htconfig.php" - changed all references from ".htconfig.php" to "/config/.htconfig.php"
- modified de, en-gb, en-us language templates
