+++ doc/Update.md
@@ -7,7 +7,7 @@ Updating Friendica
 
 If you installed Friendica in the ``path/to/friendica`` folder:
 1. Unpack the new Friendica archive in ``path/to/friendica_new``.
-2. Copy ``.htconfig.php``, ``photo/`` and ``proxy/`` from ``path/to/friendica`` to ``path/to/friendica_new``.
+2. Copy ``config/.htconfig.php``, ``photo/`` and ``proxy/`` from ``path/to/friendica`` to ``path/to/friendica_new``.
 3. Rename the ``path/to/friendica`` folder to ``path/to/friendica_old``.
 4. Rename the ``path/to/friendica_new`` folder to ``path/to/friendica``.
 5. Check your site. Note: it may go into maintenance mode to update the database schema.
