+++ src/Core/Console/NewPassword.php
@@ -58,7 +58,7 @@ HELP;
 			throw new \Asika\SimpleConsole\CommandArgsException('Too many arguments');
 		}
 
-		require_once '.htconfig.php';
+		require_once 'config/.htconfig.php';
 		$result = \dba::connect($db_host, $db_user, $db_pass, $db_data);
 		unset($db_host, $db_user, $db_pass, $db_data);
