+++ util/Doxyfile
@@ -2,7 +2,7 @@ INPUT = README.md index.php boot.php testargs.php update.php mod/ object/ includ
 RECURSIVE = YES
 PROJECT_NAME = "Friendica"
 PROJECT_LOGO = images/friendica-64.jpg
-EXCLUDE = .htconfig.php library/ doc/ .git/ log/ util/zotsh/easywebdav/ addon/ report/ privacy_image_cache/ photo/ proxy/ local/
+EXCLUDE = config/.htconfig.php library/ doc/ .git/ log/ util/zotsh/easywebdav/ addon/ report/ privacy_image_cache/ photo/ proxy/ local/
 EXCLUDE_PATTERNS = *smarty3* *strings.php*.log *.out *test* 
 OUTPUT_DIRECTORY = doc
 GENERATE_HTML = YES
