+++ view/lang/de/strings.php
@@ -809,9 +809,9 @@ $a->strings["Error: mb_string PHP module required but not installed."] = "Fehler
 $a->strings["Error: iconv PHP module required but not installed."] = "Fehler: Das iconv-Modul von PHP ist nicht installiert.";
 $a->strings["Error: POSIX PHP module required but not installed."] = "Fehler POSIX PHP Modul erforderlich aber nicht installiert.";
 $a->strings["Error, XML PHP module required but not installed."] = "Fehler: XML PHP Modul erforderlich aber nicht installiert.";
-$a->strings["The web installer needs to be able to create a file called \".htconfig.php\" in the top folder of your web server and it is unable to do so."] = "Der Installationswizard muss in der Lage sein, eine Datei im Stammverzeichnis Deines Webservers anzulegen, ist allerdings derzeit nicht in der Lage, dies zu tun.";
+$a->strings["The web installer needs to be able to create a file called \".htconfig.php\" in the \"config/\" folder of your web server and it is unable to do so."] = "Der Installationswizard muss in der Lage sein, eine Datei im Stammverzeichnis Deines Webservers anzulegen, ist allerdings derzeit nicht in der Lage, dies zu tun.";
 $a->strings["This is most often a permission setting, as the web server may not be able to write files in your folder - even if you can."] = "In den meisten Fällen ist dies ein Problem mit den Schreibrechten. Der Webserver könnte keine Schreiberlaubnis haben, selbst wenn Du sie hast.";
-$a->strings["At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica top folder."] = "Nachdem Du alles ausgefüllt hast, erhältst Du einen Text, den Du in eine Datei namens .htconfig.php in Deinem Friendica-Wurzelverzeichnis kopieren musst.";
+$a->strings["At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica \"config/\" folder."] = "Nachdem Du alles ausgefüllt hast, erhältst Du einen Text, den Du in eine Datei namens .htconfig.php in Deinem Friendica-Wurzelverzeichnis kopieren musst.";
 $a->strings["You can alternatively skip this procedure and perform a manual installation. Please see the file \"INSTALL.txt\" for instructions."] = "Alternativ kannst Du diesen Schritt aber auch überspringen und die Installation manuell durchführen. Eine Anleitung dazu (Englisch) findest Du in der Datei INSTALL.txt.";
 $a->strings[".htconfig.php is writable"] = "Schreibrechte auf .htconfig.php";
 $a->strings["Friendica uses the Smarty3 template engine to render its web views. Smarty3 compiles templates to PHP to speed up rendering."] = "Friendica nutzt die Smarty3 Template Engine um die Webansichten zu rendern. Smarty3 kompiliert Templates zu PHP um das Rendern zu beschleunigen.";
@@ -824,7 +824,7 @@ $a->strings["Url rewrite is working"] = "URL rewrite funktioniert";
 $a->strings["ImageMagick PHP extension is not installed"] = "ImageMagicx PHP Erweiterung ist nicht installiert.";
 $a->strings["ImageMagick PHP extension is installed"] = "ImageMagick PHP Erweiterung ist installiert";
 $a->strings["ImageMagick supports GIF"] = "ImageMagick unterstützt GIF";
-$a->strings["The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server root."] = "Die Konfigurationsdatei \".htconfig.php\" konnte nicht angelegt werden. Bitte verwende den angefügten Text, um die Datei im Stammverzeichnis Deiner Friendica-Installation zu erzeugen.";
+$a->strings["The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server \"config/\" directory."] = "Die Konfigurationsdatei \".htconfig.php\" konnte nicht angelegt werden. Bitte verwende den angefügten Text, um die Datei im \"config/\"-Verzeichnis Deiner Friendica-Installation zu erzeugen.";
 $a->strings["<h1>What next</h1>"] = "<h1>Wie geht es weiter?</h1>";
 $a->strings["IMPORTANT: You will need to [manually] setup a scheduled task for the worker."] = "Wichtig: Du musst [manuell] einen Cronjob (o.ä.) für den Worker einrichten.";
 $a->strings["Go to your new Friendica node <a href=\"%s/register\">registration page</a> and register as new user. Remember to use the same email you have entered as administrator email. This will allow you to enter the site admin panel."] = "Du solltest nun die Seite zur <a href=\"%s/register\">Nutzerregistrierung</a> deiner neuen Friendica Instanz besuchen und einen neuen Nutzer einrichten. Bitte denke daran die selbe E-Mail Adresse anzugeben, die du auch als Administrator E-Mail angegeben hast, damit du das Admin-Panel verwenden kannst.";
