+++ view/lang/en-us/strings.php
@@ -844,9 +844,9 @@ $a->strings["Error: The MySQL driver for PDO is not installed."] = "Error: MySQL
 $a->strings["Error: mb_string PHP module required but not installed."] = "Error: mb_string PHP module required but not installed.";
 $a->strings["Error: iconv PHP module required but not installed."] = "Error: iconv PHP module required but not installed.";
 $a->strings["Error, XML PHP module required but not installed."] = "Error, XML PHP module required but not installed.";
-$a->strings["The web installer needs to be able to create a file called \".htconfig.php\" in the top folder of your web server and it is unable to do so."] = "The web installer needs to be able to create a file called \".htconfig.php\" in the top-level directory of your web server, but it is unable to do so.";
+$a->strings["The web installer needs to be able to create a file called \".htconfig.php\" in the \"config/\" folder of your web server and it is unable to do so."] = "The web installer needs to be able to create a file called \".htconfig.php\" in the top-level directory of your web server, but it is unable to do so.";
 $a->strings["This is most often a permission setting, as the web server may not be able to write files in your folder - even if you can."] = "This is most often a permission setting issue, as the web server may not be able to write files in your directory - even if you can.";
-$a->strings["At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica top folder."] = "At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica top-level directory.";
+$a->strings["At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica \"config/\" folder."] = "At the end of this procedure, we will give you a text to save in a file named .htconfig.php in your Friendica top-level directory.";
 $a->strings["You can alternatively skip this procedure and perform a manual installation. Please see the file \"INSTALL.txt\" for instructions."] = "Alternatively, you may skip this procedure and perform a manual installation. Please see the file \"INSTALL.txt\" for instructions.";
 $a->strings[".htconfig.php is writable"] = ".htconfig.php is writable";
 $a->strings["Friendica uses the Smarty3 template engine to render its web views. Smarty3 compiles templates to PHP to speed up rendering."] = "Friendica uses the Smarty3 template engine to render its web views. Smarty3 compiles templates to PHP to speed up rendering.";
@@ -859,7 +859,7 @@ $a->strings["Url rewrite is working"] = "URL rewrite is working";
 $a->strings["ImageMagick PHP extension is not installed"] = "ImageMagick PHP extension is not installed";
 $a->strings["ImageMagick PHP extension is installed"] = "ImageMagick PHP extension is installed";
 $a->strings["ImageMagick supports GIF"] = "ImageMagick supports GIF";
-$a->strings["The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server root."] = "The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server root.";
+$a->strings["The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server \"config/\" directory."] = "The database configuration file \".htconfig.php\" could not be written. Please use the enclosed text to create a configuration file in your web server root.";
 $a->strings["<h1>What next</h1>"] = "<h1>What next</h1>";
 $a->strings["IMPORTANT: You will need to [manually] setup a scheduled task for the worker."] = "IMPORTANT: You will need to [manually] setup a scheduled task for the worker.";
 $a->strings["Go to your new Friendica node <a href=\"%s/register\">registration page</a> and register as new user. Remember to use the same email you have entered as administrator email. This will allow you to enter the site admin panel."] = "Go to your new Friendica node <a href=\"%s/register\">registration page</a> and register as new user. Remember to use the same email you have entered as administrator email. This will allow you to enter the site admin panel.";
